Use proper casts. JIRA: CXX-14376
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@106232 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/seqtable/SeqTable_single_data.cpp
@@ -186,7 +186,7 @@ void CSeqTable_single_data::GetValue(double& v) const
         v = GetInt();
         break;
     case e_Int8:
-        v = GetInt8();
+        v = static_cast<double>(GetInt8());
         break;
     case e_Real:
         v = GetReal();
